Fix using the open/close gate abilities and gates collision
No timeout, replace the ability and play the animation immediately.
Use the same trigger in the chat command triggers.
There still can be races with the chat commands and other players casting the ability.

Gates have not both abilities from the beginning which led to a bug of the first opening.

Gates have NO collision. This should fix attacking through a closed gate.
The collision of the destructable should be blocking attacks.